+++ packages/discord.js/src/client/actions/WebhooksUpdate.js
@@ -15,7 +15,7 @@ class WebhooksUpdate extends Action {
     /**
      * Emitted whenever a channel has its webhooks changed.
      * @event Client#webhooksUpdate
-     * @param {TextChannel|NewsChannel|VoiceChannel|StageChannel|ForumChannel} channel
+     * @param {TextChannel|NewsChannel|VoiceChannel|StageChannel|ForumChannel|MediaChannel} channel
      * The channel that had a webhook update
      */
     client.emit('webhooksUpdate', channel);
@@ -23,7 +23,7 @@ class WebhooksUpdate extends Action {
     /**
      * Emitted whenever a channel has its webhooks changed.
      * @event Client#webhookUpdate
-     * @param {TextChannel|NewsChannel|VoiceChannel|StageChannel|ForumChannel} channel
+     * @param {TextChannel|NewsChannel|VoiceChannel|StageChannel|ForumChannel|MediaChannel} channel
      * The channel that had a webhook update
      * @deprecated Use {@link Client#event:webhooksUpdate} instead.
      */
