+++ packages/discord.js/src/managers/GuildInviteManager.js
@@ -46,8 +46,9 @@ class GuildInviteManager extends CachedManager {
    * * NewsChannel
    * * StageChannel
    * * ForumChannel
+   * * MediaChannel
    * * Snowflake
-   * @typedef {TextChannel|VoiceChannel|NewsChannel|StageChannel|ForumChannel|Snowflake}
+   * @typedef {TextChannel|VoiceChannel|NewsChannel|StageChannel|ForumChannel|MediaChannel|Snowflake}
    * GuildInvitableChannelResolvable
    */
