+++ packages/discord.js/src/structures/AutoModerationActionExecution.js
@@ -87,7 +87,7 @@ class AutoModerationActionExecution {
 
   /**
    * The channel where this action was triggered from.
-   * @type {?(GuildTextBasedChannel|ForumChannel)}
+   * @type {?(GuildTextBasedChannel|ForumChannel|MediaChannel)}
    * @readonly
    */
   get channel() {
