+++ packages/discord.js/src/structures/MediaChannel.js
@@ -1,0 +1,11 @@
+'use strict';
+
+const ThreadOnlyChannel = require('./ThreadOnlyChannel');
+
+/**
+ * Represents a media channel.
+ * @extends {ThreadOnlyChannel}
+ */
+class MediaChannel extends ThreadOnlyChannel {}
+
+module.exports = MediaChannel;
